undo refactor styles
